reorganize asi stages
